move secondq to algo
