+++ qc2/algorithms/pennylane/qpe/pebase.py
@@ -4,14 +4,13 @@ import pennylane as qml
 from pennylane import numpy as np
 from pennylane import QNode
 from pennylane.operation import Operator
-from qc2.second_q.active_space import ActiveSpace
 from qc2.algorithms.results import QPEResults
 from qc2.algorithms.base.qc2_algorithm_base_class import BaseAlgorithm
 from qc2.algorithms.pennylane.qubit_mappers.jordan_wigner import JordanWigner
 
 
 from qc2.qc2_driver import QC2
-from qc2.second_q.active_space import ActiveSpace
+from qc2.algorithms.second_q.active_space import ActiveSpace
 from qc2.algorithms.pennylane.qubit_mappers.base_mapper import PennylaneBaseMapper
