+++ qc2/algorithms/pennylane/qpe/qpe.py
@@ -6,7 +6,7 @@ from pennylane.operation import Operator
 
 
 from qc2.qc2_driver import QC2
-from qc2.second_q.active_space import ActiveSpace
+from qc2.algorithms.second_q.active_space import ActiveSpace
 from qc2.algorithms.pennylane.qubit_mappers.base_mapper import PennylaneBaseMapper
 
 from .pebase import PEBase
