+++ qc2/algorithms/pennylane/vqe/oo_vqe.py
@@ -4,7 +4,7 @@ from qc2.algorithms.pennylane.vqe.sa_oo_vqe import SA_OO_VQE
 from qc2.algorithms.pennylane.ansatz.generate_ansatz import generate_ansatz
 
 from qc2.qc2_driver import QC2
-from qc2.second_q.active_space import ActiveSpace
+from qc2.algorithms.second_q.active_space import ActiveSpace
 from qc2.algorithms.pennylane.qubit_mappers.base_mapper import PennylaneBaseMapper
 
 class OO_VQE(SA_OO_VQE):
