+++ qc2/algorithms/qiskit/qpe/iqpe.py
@@ -6,7 +6,7 @@ from qiskit.primitives import BaseSampler
 from .pebase import PEBase
 
 from qc2.qc2_driver import QC2
-from qc2.second_q.active_space import ActiveSpace
+from qc2.algorithms.second_q.active_space import ActiveSpace
 from qc2.algorithms.qiskit.qubit_mappers.base_mapper import QiskitBaseMapper
 
 class QC2IterativePhaseEstimation(IterativePhaseEstimation):
