+++ qc2/algorithms/qiskit/qpe/pebase.py
@@ -6,12 +6,11 @@ from qiskit.primitives import Sampler
 from qiskit.circuit.library import UnitaryGate
 
 from qc2.algorithms.base.qc2_algorithm_base_class import BaseAlgorithm
-from qc2.second_q.active_space import ActiveSpace
 
 from qc2.algorithms.results import QPEResults
 from qc2.algorithms.qiskit.qubit_mappers.jordan_wigner import JordanWigner
 from qc2.qc2_driver import QC2
-from qc2.second_q.active_space import ActiveSpace
+from qc2.algorithms.second_q.active_space import ActiveSpace
 from qc2.algorithms.qiskit.qubit_mappers.base_mapper import QiskitBaseMapper
 from qc2.algorithms.qiskit.ansatz.hatree_fock import HartreeFock
