+++ qc2/algorithms/qiskit/vqe/sa_oo_vqe.py
@@ -8,12 +8,12 @@ from functools import partial
 
 
 from qc2.algorithms.qiskit.ansatz.ucc import UCC
-from qc2.second_q.fermionic_operator import FermionicOperator
+from qc2.algorithms.second_q.fermionic_operator import FermionicOperator
 from qc2.algorithms.qiskit.qubit_mappers.base_mapper import QiskitBaseMapper
 from qc2.qc2_driver import QC2
 from qc2.algorithms.qiskit.vqe.vqe import VQE
 from qc2.algorithms.results import SAOOVQEResults
-from qc2.second_q.active_space import ActiveSpace
+from qc2.algorithms.second_q.active_space import ActiveSpace
 from qc2.algorithms.utils.orbital_optimization import OrbitalOptimization
 from qc2.algorithms.qiskit.ansatz.state_resolution import StateResolution
 from qc2.algorithms.qiskit.ansatz.generate_ansatz import generate_ansatz
