+++ qc2/algorithms/second_q/fermionic_operator.py
