+++ tests/integration_tests/test_vqe_qiskit_triplet_carbon.py
@@ -9,7 +9,7 @@ from qiskit.primitives import Estimator
 from qc2.qc2_driver import QC2 as qc2Data
 from qc2.ase import PySCF
 from qc2.algorithms.qiskit import VQE
-from qc2.second_q.active_space import ActiveSpace
+from qc2.algorithms.second_q.active_space import ActiveSpace
 from qc2.algorithms.qiskit.qubit_mappers.jordan_wigner import JordanWigner
