+++ tests/unit_tests/test_pennylane_vqe_class.py
@@ -6,7 +6,7 @@ from ase.build import molecule
 
 from qc2.qc2_driver import QC2 as qc2Data
 from qc2.ase import PySCF
-from qc2.second_q.active_space import ActiveSpace
+from qc2.algorithms.second_q.active_space import ActiveSpace
 
 
 try:
